+++ test-resources/ping_imp.clj
@@ -1,0 +1,3 @@
+(ns ping-imp)
+
+(defn pong [] "pong-deferred")